requirements.txt
certifi
charset-normalizer
idna
PyJWT
python-dateutil
python-dotenv
requests
six
sqlparse
urllib3
